+++ .gitignore
@@ -57,4 +57,5 @@ spu_wheel.name
 
 /pass-test/reduce_case
 /pass-test/reduce_log
+/pass-test/test-log
 /xla
